Rename pkg/networkservicemesh to pkg/networkservice
The goal had been to name the top level directories thematically under pkg/
mostly after their APIs, so:

pkg/networkservice
pkg/registry
etc

This is the correction.
