+++ .golangci.yml
@@ -159,23 +159,23 @@ issues:
   exclude-rules:
     # We really *do* want to pass a pointer to an interface in these cases.  See comments in file for New{Server,Client}
     # function for why
-    - path: pkg/networkservicemesh/common/monitor/server.go
+    - path: pkg/networkservice/common/monitor/server.go
       linters:
         - gocritic
       text: "ptrToRefParam"
-    - path: pkg/networkservicemesh/common/timeout/server.go
+    - path: pkg/networkservice/common/timeout/server.go
       linters:
         - gocritic
       text: "ptrToRefParam"
-    - path: pkg/networkservicemesh/common/heal/client.go
+    - path: pkg/networkservice/common/heal/client.go
       linters:
         - gocritic
       text: "ptrToRefParam"
-    - path: pkg/networkservicemesh/chains/client/client.go
+    - path: pkg/networkservice/chains/client/client.go
       linters:
         - gocritic
       text: "ptrToRefParam"
-    - path: pkg/networkservicemesh/common/localbypass/networkservice_server.go
+    - path: pkg/networkservice/common/localbypass/networkservice_server.go
       linters:
         - gocritic
       text: "ptrToRefParam"
