+++ pkg/networkservice/common/authorize/README.md
