+++ pkg/networkservice/common/discover/context.go
