+++ pkg/networkservice/core/adapters/client_to_server.go
