+++ pkg/networkservice/core/adapters/server_to_client.go
