--- pkg/networkservicemesh/common/clienturl/server.go
@@ -1,48 +1,0 @@
-// Copyright (c) 2020 Cisco Systems, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package clienturl
-
-import (
-	"context"
-	"net/url"
-
-	"github.com/golang/protobuf/ptypes/empty"
-
-	"github.com/networkservicemesh/networkservicemesh/controlplane/api/connection"
-	"github.com/networkservicemesh/networkservicemesh/controlplane/api/networkservice"
-
-	"github.com/networkservicemesh/sdk/pkg/networkservicemesh/core/next"
-)
-
-type clientURL struct {
-	u *url.URL
-}
-
-// NewServer - returns a new server that will set the request context to return url u when ClientURL(ctx) is called
-func NewServer(u *url.URL) networkservice.NetworkServiceServer {
-	return &clientURL{u: u}
-}
-
-func (c *clientURL) Request(ctx context.Context, request *networkservice.NetworkServiceRequest) (*connection.Connection, error) {
-	ctx = WithClientURL(ctx, c.u)
-	return next.Server(ctx).Request(ctx, request)
-}
-
-func (c *clientURL) Close(ctx context.Context, conn *connection.Connection) (*empty.Empty, error) {
-	ctx = WithClientURL(ctx, c.u)
-	return next.Server(ctx).Close(ctx, conn)
-}
